fix(64-bit): widen truncating pointer casts (PushBuffer)

+++ game/PushBuffer.c
@@ -208,7 +208,7 @@ void PushBuffer_SetDrawEnv_DecalMP(void *ot, struct DB *backBuffer, RECT *viewpo
 	if (p <= backBuffer->primMem.guardEnd)
 	{
 		// advance curr
-		backBuffer->primMem.cursor = (void *)((u32)backBuffer->primMem.cursor + 0x40);
+		backBuffer->primMem.cursor = (void *)((uintptr_t)backBuffer->primMem.cursor + 0x40);
 
 		prim = p;
 	}
@@ -267,7 +267,7 @@ void PushBuffer_SetDrawEnv_Normal(void *ot, struct PushBuffer *pb, struct DB *ba
 	void *p = backBuffer->primMem.cursor;
 	if (p <= backBuffer->primMem.guardEnd)
 	{
-		backBuffer->primMem.cursor = (void *)((u32)backBuffer->primMem.cursor + 0x40);
+		backBuffer->primMem.cursor = (void *)((uintptr_t)backBuffer->primMem.cursor + 0x40);
 
 		SetDrawEnv(p, &newDrawEnv);
 
@@ -340,11 +340,11 @@ void PushBuffer_SetMatrixVP(struct PushBuffer *pb)
 	sVar7 = *(s16 *)&matrixDST->m[2][2];
 
 	// CameraMatrix, for shadows, particles, and audio
-	*(int *)((int)&pb->matrix_Camera + 0x0) = uVar3;
-	*(int *)((int)&pb->matrix_Camera + 0x4) = uVar4;
-	*(int *)((int)&pb->matrix_Camera + 0x8) = uVar5;
-	*(int *)((int)&pb->matrix_Camera + 0xC) = uVar6;
-	*(s16 *)((int)&pb->matrix_Camera + 0x10) = sVar7;
+	*(int *)((uintptr_t)&pb->matrix_Camera + 0x0) = uVar3;
+	*(int *)((uintptr_t)&pb->matrix_Camera + 0x4) = uVar4;
+	*(int *)((uintptr_t)&pb->matrix_Camera + 0x8) = uVar5;
+	*(int *)((uintptr_t)&pb->matrix_Camera + 0xC) = uVar6;
+	*(s16 *)((uintptr_t)&pb->matrix_Camera + 0x10) = sVar7;
 
 	// transpose the camera matrix
 	view0 = uVar3 & 0xffff | uVar4 & 0xffff0000;
@@ -353,11 +353,11 @@ void PushBuffer_SetMatrixVP(struct PushBuffer *pb)
 	viewC = uVar4 & 0xffff | uVar5 & 0xffff0000;
 
 	// CameraTranspose, for lightning during Driver Warping effect
-	*(int *)((int)&pb->matrix_CameraTranspose + 0x0) = view0;
-	*(int *)((int)&pb->matrix_CameraTranspose + 0x4) = view4;
-	*(int *)((int)&pb->matrix_CameraTranspose + 0x8) = view8;
-	*(int *)((int)&pb->matrix_CameraTranspose + 0xC) = viewC;
-	*(s16 *)((int)&pb->matrix_CameraTranspose + 0x10) = sVar7;
+	*(int *)((uintptr_t)&pb->matrix_CameraTranspose + 0x0) = view0;
+	*(int *)((uintptr_t)&pb->matrix_CameraTranspose + 0x4) = view4;
+	*(int *)((uintptr_t)&pb->matrix_CameraTranspose + 0x8) = view8;
+	*(int *)((uintptr_t)&pb->matrix_CameraTranspose + 0xC) = viewC;
+	*(s16 *)((uintptr_t)&pb->matrix_CameraTranspose + 0x10) = sVar7;
 
 	// load transpose camera matrix
 	// similar to gte_SetLightMatrix
@@ -379,11 +379,11 @@ void PushBuffer_SetMatrixVP(struct PushBuffer *pb)
 	CTR_GteStoreMAC(&pb->matrix_ViewProj.t[0]);
 
 	// start with transpose camera matrix
-	*(int *)((int)&pb->matrix_ViewProj + 0x0) = view0;
-	*(int *)((int)&pb->matrix_ViewProj + 0x4) = view4;
-	*(int *)((int)&pb->matrix_ViewProj + 0x8) = view8;
-	*(int *)((int)&pb->matrix_ViewProj + 0xC) = viewC;
-	*(s16 *)((int)&pb->matrix_ViewProj + 0x10) = sVar7;
+	*(int *)((uintptr_t)&pb->matrix_ViewProj + 0x0) = view0;
+	*(int *)((uintptr_t)&pb->matrix_ViewProj + 0x4) = view4;
+	*(int *)((uintptr_t)&pb->matrix_ViewProj + 0x8) = view8;
+	*(int *)((uintptr_t)&pb->matrix_ViewProj + 0xC) = viewC;
+	*(s16 *)((uintptr_t)&pb->matrix_ViewProj + 0x10) = sVar7;
 
 	// NTSC:
 	// 0x360/0x600 = 9/16 aspect,
